requirements.txt
selene~=2.0.0rc9
pytest~=8.3.3
Appium-Python-Client~=4.2.0
python-dotenv~=1.0.1
selenium~=4.25.0
allure-python-commons~=2.13.5
allure-pytest == 2.13.5
requests~=2.32.3